% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills.R
\name{skill_entries}
\alias{skill_entries}
\title{Create skill entries}
\usage{
skill_entries(data, type, details, .protect = TRUE)
}
\arguments{
\item{type}{Skill type/category}

\item{details}{Skill details/descriptions}
}
\value{
Object of class \code{vitae_skill}
}
\description{
Create skill entries
}
